+++ core/player.py
@@ -1,11 +1,16 @@
 """Modulo Player.
 Define la clase Player que representa a un jugador del juego."""
+from core.redis_store import RedisStore
 class Player:
     """La clase player define a cada uno de los dos jugadores
     junto con sus responsabilidades. Atributos: name, checker_type, bar_index"""
     def __init__(self, name:str, checker_type:int):
-        self.__name__ = name
+        self.__redis_store__ = RedisStore()
         self.__checker_type__ = checker_type
+        if not self.__redis_store__.get_value(f'name{checker_type}'):
+            self.__name__ = name
+        else:
+            self.__name__ = self.__redis_store__.get_value(f'name{checker_type}')
         self.__bar_index__ = 24 if checker_type == 1 else 25
     def get_name(self):
         """Devuelve el nombre del jugador (str)."""